Asynchronous pipe using libuv

+++ benc/String.c
@@ -28,7 +28,7 @@ String* String_new(const char* bytes, struct Allocator* allocator)
 }
 
 /** @see Object.h */
-String* String_newBinary(const char* bytes, size_t length, struct Allocator* allocator)
+String* String_newBinary(const char* bytes, unsigned long length, struct Allocator* allocator)
 {
     char* copy = Allocator_malloc(allocator, length + 1);
     // Make the string null terminated so it will print nicely.
